requirements.txt
ttkbootstrap>=1.10.1
toml
Pillow
eyeD3>=0.9.7